FEBio3: Modified surface elements so they can support mixed formulations.
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/FEBio3@12944 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ FEBio3/FEBioCommand.cpp
@@ -709,7 +709,7 @@ int FEBioCmd_list::run(int nargs, char** argv)
 		{
 			if (nargs <= nc + 1) return invalid_nr_args();
 			szpat = argv[++nc];
-			lpat = strlen(szpat);
+			lpat = (int)strlen(szpat);
 			if (lpat == 0) szpat = 0;
 		}
 		else return unknown_args();
